Reorganize specs
