+++ spec/game.specs/player.spec.js
@@ -8,9 +8,9 @@
 
 const util = require('util')
 
-const Player = require('../lib/game/player')
-const Vector2D = require('../lib/game/physics/vector-2d')
-const BoundEntity = require('../lib/game/physics/bound-entity')
+const Player = require('../../lib/game/player')
+const Vector2D = require('../../lib/game/physics/vector-2d')
+const BoundEntity = require('../../lib/game/physics/bound-entity')
 
 const playerStats = {
   name: 'THISISTHEFBI',
